backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
langchain
openai
faiss-cpu
python-dotenv
black
flake8
pytest
nltk
python-jose[cryptography]
passlib[bcrypt]
langchain-community
langchain_openai
weaviate-client>=4.0.0
